refactor: extract assets feature folder
Move Asset schemas, video-conversion, image-decorations,
image-autocomplete, asset handler factories (from editor-utils),
and sidebar-assets into src/app/features/assets/. Expose an
imageExtensions factory that plugs into the editor extension slot.
