+++ src/app/features/assets/lib/image-decorations.ts
