+++ src/app/features/assets/widgets/sidebar-assets.tsx
