+++ src/app/features/import-export/widgets/sidebar-import-export.tsx
@@ -2,7 +2,7 @@ import { useRef, useState } from "react"
 import { Group, co } from "jazz-tools"
 import { createImage } from "jazz-tools/media"
 import { Document, Asset, ImageAsset, VideoAsset } from "@/schema"
-import { compressVideo, canEncodeVideo } from "@/lib/video-conversion"
+import { compressVideo, canEncodeVideo } from "@/app/features/assets"
 import { getDocumentTitle } from "@/lib/document-utils"
 import { getPath } from "@/editor/frontmatter"
 import { Button } from "@/components/ui/button"
