+++ src/app/routes/doc.$id.index.tsx
@@ -11,6 +11,12 @@ import { useCoState, useAccount, useIsAuthenticated } from "jazz-tools/react"
 import { Document, Space, UserAccount, createSpaceDocument } from "@/schema"
 import { z } from "zod"
 import {
+	handleSaveCopy,
+	setupKeyboardShortcuts,
+	resolve,
+	type LoadedDocument,
+} from "@/lib/editor-utils"
+import {
 	makeUploadImage,
 	makeUploadVideo,
 	makeUploadAssets,
@@ -18,12 +24,11 @@ import {
 	makeIsAssetUsed,
 	makeDeleteAsset,
 	makeDownloadAsset,
-	handleSaveCopy,
-	setupKeyboardShortcuts,
-	resolve,
 	canEncodeVideo,
-	type LoadedDocument,
-} from "@/lib/editor-utils"
+	imageExtensions,
+	SidebarAssets,
+	type SidebarAsset,
+} from "@/app/features/assets"
 import {
 	MarkdownEditor,
 	useMarkdownEditorRef,
@@ -69,7 +74,6 @@ import { SidebarFileMenu } from "@/components/sidebar-file-menu"
 import { SidebarEditMenu } from "@/components/sidebar-edit-menu"
 import { SidebarFormatMenu } from "@/components/sidebar-format-menu"
 import { SidebarCollaboration } from "@/app/features/sharing"
-import { SidebarAssets, type SidebarAsset } from "@/components/sidebar-assets"
 import { ThemeToggle, useTheme } from "@/app/components/appearance"
 import {
 	Tooltip,
@@ -262,6 +266,10 @@ function EditorContent({ doc, docId }: EditorContentProps) {
 			video: a.type === "video" ? a.video : undefined,
 			muteAudio: a.type === "video" ? a.muteAudio : undefined,
 		})) ?? []
+	let assetsRef = useRef(assets)
+	useEffect(() => {
+		assetsRef.current = assets
+	})
 
 	// Get documents for wikilink autocomplete - personal docs only
 	let documents: WikilinkDoc[] = []
@@ -492,7 +500,19 @@ function EditorContent({ doc, docId }: EditorContentProps) {
 					onUploadImage={makeUploadImage(doc)}
 					onUploadVideo={canUploadVideo ? makeUploadVideo(doc) : undefined}
 					autoSortTasks={editorSettings?.editor?.autoSortTasks}
-					extensions={[]}
+					extensions={[
+						imageExtensions({
+							resolver: assetId => {
+								let asset = assetsRef.current.find(a => a.id === assetId)
+								if (!asset) return undefined
+								return { url: `asset:${assetId}`, type: asset.type }
+							},
+							onPreview: (url, alt) =>
+								editor.current?.showImagePreview(url, alt),
+							getAssets: () =>
+								assetsRef.current.map(a => ({ id: a.id, name: a.name })),
+						}),
+					]}
 				/>
 				<EditorToolbar
 					editor={editor}
